[DriverPersonaBO.java] Added parameter option to enable and set blacklist for driver names
This feature allows server owners to enable and set a blacklist of driver persona names from server parameter that will prevent users who create new drivers to use them (will notify the user that the name is taken even though it really isn't).

A good few examples of those names are: ADMIN,DEVELOPER,MODERATOR,SERVER,SYSTEM

Tested and working properly on Underground Stage server

UPDATE 3: Corrected a burnt-out typo of mine and revered back to the .equals() way since we do not want people with substrings of the blacklisted names (such as ADMI) to also be accounted. Added an extra return of arrayOfString from the blacklist check statement so the code will not execute an extra query if the requested driver name is blacklisted